+++ src/gql/relay/mixins/with_order_by_filter.mixin.ts
@@ -1,8 +1,8 @@
 import { ArgsType, Field } from 'type-graphql';
-import { RelayOrderByGQL, RelayOrderBy } from '../scalars/relay-order-by.scalar';
-import { ClassType } from '../../helpers';
+import { RelayOrderByGQL, RelayOrderBy } from '../../scalars/relay-order-by.scalar';
+import { ClassType } from '../../../helpers';
 
-export function withOrderByFilter<TFilterBase extends ClassType<any>>(Base: TFilterBase) {
+export function withOrderByFilter<TFilterBase extends ClassType<object>>(Base: TFilterBase) {
   @ArgsType()
   class OrderByFilter extends Base {
     @Field(type => RelayOrderByGQL, { nullable: true })