+++ src/gql/relay/mixins/with_pagination_filter.mixin.ts
@@ -1,8 +1,8 @@
-import { ClassType } from '../../helpers/classtype';
 import { ArgsType, Field, Int, ID } from 'type-graphql';
 import { Min } from 'class-validator';
+import { ClassType } from '../../../helpers/classtype';
 
-export function withPaginationFilter<TFilterBase extends ClassType<any>>(Base: TFilterBase) {
+export function withPaginationFilter<TFilterBase extends ClassType<object>>(Base: TFilterBase) {
   @ArgsType()
   class ConnectionFilter extends Base {
     @Field(type => Int, { nullable: true })